Refactored code for calling a compound task
Lifted it out of Call.CallTask and put it in CompoundTask.Call, then modified Module.Call* to use that.

That required correcting a bunch of tests that hadn't bothered to declare fail tests to be [fallible].

This adds an allocation and call of a new continuation to each CompoundTask call but it's worth it for code simplication, since there had by this point been several places that had either called a compound task or would have liked to have done so.