BREAKING CHANGE: split Not into Not and NotAny
Because the semantics of negation by failure means that [Not [P ?x]] means "P is false for all ?x", not "find me an ?x for which P is false", Not causes a lot of subtle bugs.  So now Not has been modified to require its goal be ground (no variables).  The functionality of the original version is available under the name NotAny.